requirements.txt
aggdraw==1.3.19
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
banks==2.1.2
bayesian-optimization==1.4.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
botorch==0.11.1
braceexpand==0.1.7
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
cma==4.0.0
colorama==0.4.6
confection==0.1.5
conllu==6.0.0
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
deap==1.4.2
Deprecated==1.2.15
dgl==2.1.0
directsearch==1.0
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
elpv-dataset==1.0.0.post1
en_core_sci_lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz#sha256=a2bfde10824958d51cd016b4142563373d553e24767d79b1abf8699596e31a5b
en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz#sha256=7c8fc52542dd1452ffce00b045c1298e2c185b7cf84793f8e0ec941987c09808
en_core_sci_scibert @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz#sha256=be4afaa2710d63206fafa7864001acbfab1225800bbab577f32765435e29b7f2
en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz#sha256=a54a114ee11dab7cf6db8cf69ef485145011537f45f3c16c71338c48fd6f070b
en_ner_bc5cdr_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz#sha256=ffc73130a710edf851206199720cb2c744a043e032f5da6ba4bb36863deca778
en_ner_bionlp13cg_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bionlp13cg_md-0.5.4.tar.gz#sha256=d747ec2b3871d7f9187d4f4769e3f7014a48e804b01c18e3da9991a67024eae1
en_ner_craft_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_craft_md-0.5.4.tar.gz#sha256=63ddbc8320c968e419148e4c10dad595cc601aadc6f35187a1340d147f706ab7
en_ner_jnlpba_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_jnlpba_md-0.5.4.tar.gz#sha256=9a6f59629c1aa0d43643d3e0f786016766149b2ef622d4d7a08808c4a6d9cb11
et_xmlfile==2.0.0
evotorch==0.5.1
Farama-Notifications==0.0.4
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
GPUtil==1.4.0
gpytorch==1.11
graphviz==0.20.3
greenlet==3.2.2
griffe==1.7.3
gymnasium==1.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.7
imageio==2.37.0
imbalanced-learn==0.12.3
imblearn==0.0
instructor==1.3.4
jaxtyping==0.2.33
Jinja2==3.1.4
jiter==0.4.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
libauc==1.4.0
lightgbm==4.5.0
lightning-utilities==0.11.6
linear-operator==0.5.1
littleutils==0.2.4
llama-index-core==0.12.37
llama-index-llms-ollama==0.5.4
llama-index-tools-mcp==0.1.3
loguru==0.7.2
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.0
mcp==1.9.0
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multipledispatch==1.0.0
munkres==1.1.4
murmurhash==1.0.11
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.3
nevergrad==1.0.12
nflgame==1.2.20
nflgame-redux==3.0.0
nltk==3.9.1
nmslib-metabrainz==2.1.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
ollama==0.4.5
openai==1.81.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
outdated==0.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.8
plotly==5.22.0
preshed==3.0.9
protobuf==5.29.3
psutil==6.0.0
pyarrow==18.1.0
pybaseball==2.2.7
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyg-lib==0.4.0+pt23cpu
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.9.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-ignite==0.5.0.post2
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.2
qhoptim==1.1.0
ray==2.40.0
rdkit==2024.3.5
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpds-py==0.22.3
ruff==0.11.9
safetensors==0.5.2
scikit-learn==1.5.0
scipy==1.14.0
scispacy==0.5.5
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
SQLAlchemy==2.0.41
srsly==2.5.0
sse-starlette==2.3.5
starlette==0.46.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.3.1
torch-geometric==2.6.1
torch-optimizer==0.3.0
torch_cluster==1.6.3+pt23cpu
torch_scatter==2.1.2+pt23cpu
torch_sparse==0.6.18+pt23cpu
torch_spline_conv==1.2.2+pt23cpu
torchaudio==2.3.1
torchdata==0.8.0
torchkge==0.17.7
torchmetrics==1.4.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.36.2
triton==2.3.1
typeguard==2.13.3
typer==0.12.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
visualkeras==0.1.4
wasabi==1.1.3
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
wrapt==1.17.1
xgboost==2.1.3
yarl==1.9.4
